1차 카테고리, 2차 카테고리로 상품 조회 기능 추가, 상품 ID로 상품 조회 기능 추가, 상품 조회 시 카테고리 창 변경
