+++ routes/index.js
@@ -2,14 +2,12 @@ const express = require('express');
 const middleware = require('../middlewares/middleware');
 const userRouter = require('./user');
 const productRouter = require('./product.route');
-const filterRouter = require('./filter');
 // const otherRouter = require('./other');
 
 const router = express.Router();
 
 router.use(userRouter);
 router.use('/product', productRouter);
-router.use(filterRouter);
 router.use(middleware.errorHandler);
 // router.use(otherRouter);
