+++ routes/product.route.js
@@ -1,10 +1,12 @@
 const express = require('express');
 const productCtl = require('../controllers/product.controller');
-const {asyncWrap} = require('../utils/myutils');
+const { asyncWrap } = require('../utils/myutils');
 
 const router = express.Router();
 
 router.get('/test', asyncWrap(productCtl.test));
+router.get('/filter', asyncWrap(productCtl.filter));
+router.get('/:Id', asyncWrap(productCtl.search));
 
 /*
   {
